+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Features:
   + Added "auxiliary sources" option to support IDL sources that are loaded into the model but are
     not generated from.
+- Breaking changes:
+  + Removed "stdout" command line option.
 
 ## 5.8.1
 - Release date: 2019-11-13